+++ wies/core/editables/__init__.py
@@ -2,11 +2,11 @@
 
 from typing import TYPE_CHECKING
 
-from wies.core.inline_edit.editables.assignment import AssignmentEditables
-from wies.core.inline_edit.editables.colleague import ColleagueEditables
-from wies.core.inline_edit.editables.placement import PlacementEditables
-from wies.core.inline_edit.editables.service import ServiceEditables
-from wies.core.inline_edit.editables.user_profile import UserProfileEditables
+from wies.core.editables.assignment import AssignmentEditables
+from wies.core.editables.colleague import ColleagueEditables
+from wies.core.editables.placement import PlacementEditables
+from wies.core.editables.service import ServiceEditables
+from wies.core.editables.user_profile import UserEditables
 
 if TYPE_CHECKING:
     from wies.core.inline_edit.base import EditableSet
@@ -16,7 +16,7 @@ REGISTRY: dict[str, type[EditableSet]] = {
     "colleague": ColleagueEditables,
     "placement": PlacementEditables,
     "service": ServiceEditables,
-    "user": UserProfileEditables,
+    "user": UserEditables,
 }
 
 __all__ = [
@@ -25,5 +25,5 @@ __all__ = [
     "ColleagueEditables",
     "PlacementEditables",
     "ServiceEditables",
-    "UserProfileEditables",
+    "UserEditables",
 ]